refactor: add ContinueWatching entity and conversion functions for data transfer
